requirements.txt
asyncpg==0.22.0
certifi==2020.12.5
chardet==4.0.0
freezegun==1.1.0
idna==2.10
kafka-python==2.0.2
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
